console: Rename bulk creation to device import
