+++ pkg/webui/console/views/device-add/device-add-single.styl
