+++ pkg/webui/console/views/devices/index.js
@@ -16,8 +16,8 @@ import React from 'react'
 import { Switch, Route } from 'react-router-dom'
 
 import DeviceList from '../device-list'
-import DeviceAdd from '../device-add-single'
-import DeviceAddBulk from '../device-add-bulk'
+import DeviceAdd from '../device-add'
+import DeviceImport from '../device-import'
 import Device from '../device'
 
 import sharedMessages from '../../../lib/shared-messages'
@@ -34,7 +34,7 @@ export default class Devices extends React.Component {
     return (
       <Switch>
         <Route path={`${path}/add`} component={DeviceAdd} />
-        <Route path={`${path}/import`} component={DeviceAddBulk} />
+        <Route path={`${path}/import`} component={DeviceImport} />
         <Route path={`${path}/:devId`} component={Device} />
         <Route path={`${path}`} component={DeviceList} />
       </Switch>
