+++ pkg/webui/lib/shared-messages.js
@@ -42,19 +42,17 @@ export default defineMessages({
   approve: 'Approve',
   appSKey: 'AppSKey',
   brand: 'Brand',
-  bulk: 'Bulk',
-  bulkCreation: 'Bulk Creation',
   cancel: 'Cancel',
   changeLocation: 'Change location settings',
   changePassword: 'Change Password',
   clear: 'Clear',
+  collaborator: 'Collaborator',
   collaboratorDeleteSuccess: 'Successfully removed collaborator',
   collaboratorEdit: 'Edit {collaboratorId}',
   collaboratorEditRights: 'Edit rights of {collaboratorId}',
   collaboratorId: 'Collaborator ID',
   collaboratorModalWarning: 'Are you sure you want to remove {collaboratorId} as a collaborator?',
   collaboratorRemove: 'Collaborator Remove',
-  collaborator: 'Collaborator',
   collaborators: 'Collaborators',
   collaboratorUpdateSuccess: 'Successfully updated collaborator rights',
   componentApplicationServer: 'Application Server',
@@ -106,6 +104,8 @@ export default defineMessages({
   hardware: 'Hardware',
   hardwareVersion: 'Hardware version',
   id: 'ID',
+  import: 'Import',
+  importDevices: 'Import Devices',
   integrations: 'Integrations',
   joinAccept: 'Join Accept',
   joinEUI: 'JoinEUI',
