+++ pkg/webui/locales/en.json
@@ -37,10 +37,6 @@
   "console.components.api-key-modal.index.buttonMessage": "I have copied the key",
   "console.components.api-key-modal.index.grantedRights": "Granted Rights",
   "console.components.api-key-modal.index.description": "Your API Key has been created successfully.\nNote: After closing this window, the value of the key secret will not be accessible anymore.\nMake sure to copy and store it in a safe place now.",
-  "console.components.device-bulk-create-form.index.fileImport": "File Import",
-  "console.components.device-bulk-create-form.index.file": "File",
-  "console.components.device-bulk-create-form.index.createDevices": "Create Devices",
-  "console.components.device-bulk-create-form.index.selectAFile": "Please select a template file",
   "console.components.device-data-form.messages.abp": "Activation By Personalization (ABP)",
   "console.components.device-data-form.messages.activationMode": "Activation Mode",
   "console.components.device-data-form.messages.activationSettings": "Activation Settings",
@@ -73,6 +69,10 @@
   "console.components.device-data-form.messages.validate16": "This value needs to be exactly 16 characters long",
   "console.components.device-data-form.messages.validate32": "This value needs to be exactly 32 characters long",
   "console.components.device-data-form.messages.validate8": "This value needs to be exactly 8 characters long",
+  "console.components.device-import-form.index.fileImport": "File Import",
+  "console.components.device-import-form.index.file": "File",
+  "console.components.device-import-form.index.createDevices": "Create Devices",
+  "console.components.device-import-form.index.selectAFile": "Please select a template file",
   "console.components.gateway-data-form.index.enforced": "Enforced",
   "console.components.gateway-data-form.index.dutyCycle": "Duty Cycle",
   "console.components.gateway-data-form.index.gatewayIdPlaceholder": "my-new-gateway",
@@ -131,14 +131,14 @@
   "console.containers.collaborators-table.index.id": "User / Organization ID",
   "console.containers.dev-addr-input.dev-addr-input.generate": "Generate Device Address",
   "console.containers.dev-addr-input.index.devAddrFetchingFailure": "Could not generate device address",
-  "console.containers.device-bulk-creator.index.proceed": "Proceed",
-  "console.containers.device-bulk-creator.index.retry": "Retry",
-  "console.containers.device-bulk-creator.index.converting": "Converting Templates…",
-  "console.containers.device-bulk-creator.index.creating": "Creating devices…",
-  "console.containers.device-bulk-creator.index.operationInProgress": "Operation in progress",
-  "console.containers.device-bulk-creator.index.operationHalted": "Operation halted",
-  "console.containers.device-bulk-creator.index.operationFinished": "Operation finished",
-  "console.containers.device-bulk-creator.index.errorTitle": "Could not complete operation",
+  "console.containers.device-importer.index.proceed": "Proceed",
+  "console.containers.device-importer.index.retry": "Retry",
+  "console.containers.device-importer.index.converting": "Converting Templates…",
+  "console.containers.device-importer.index.creating": "Creating devices…",
+  "console.containers.device-importer.index.operationInProgress": "Operation in progress",
+  "console.containers.device-importer.index.operationHalted": "Operation halted",
+  "console.containers.device-importer.index.operationFinished": "Operation finished",
+  "console.containers.device-importer.index.errorTitle": "Could not complete operation",
   "console.containers.device-template-format-select.index.title": "Format",
   "console.containers.device-template-format-select.index.warning": "Could not retrieve the list of available device template formats",
   "console.containers.freq-plans-select.index.title": "Frequency Plan",
@@ -266,19 +266,17 @@
   "lib.shared-messages.approve": "Approve",
   "lib.shared-messages.appSKey": "AppSKey",
   "lib.shared-messages.brand": "Brand",
-  "lib.shared-messages.bulk": "Bulk",
-  "lib.shared-messages.bulkCreation": "Bulk Creation",
   "lib.shared-messages.cancel": "Cancel",
   "lib.shared-messages.changeLocation": "Change location settings",
   "lib.shared-messages.changePassword": "Change Password",
   "lib.shared-messages.clear": "Clear",
+  "lib.shared-messages.collaborator": "Collaborator",
   "lib.shared-messages.collaboratorDeleteSuccess": "Successfully removed collaborator",
   "lib.shared-messages.collaboratorEdit": "Edit {collaboratorId}",
   "lib.shared-messages.collaboratorEditRights": "Edit rights of {collaboratorId}",
   "lib.shared-messages.collaboratorId": "Collaborator ID",
   "lib.shared-messages.collaboratorModalWarning": "Are you sure you want to remove {collaboratorId} as a collaborator?",
   "lib.shared-messages.collaboratorRemove": "Collaborator Remove",
-  "lib.shared-messages.collaborator": "Collaborator",
   "lib.shared-messages.collaborators": "Collaborators",
   "lib.shared-messages.collaboratorUpdateSuccess": "Successfully updated collaborator rights",
   "lib.shared-messages.componentApplicationServer": "Application Server",
@@ -330,6 +328,8 @@
   "lib.shared-messages.hardware": "Hardware",
   "lib.shared-messages.hardwareVersion": "Hardware version",
   "lib.shared-messages.id": "ID",
+  "lib.shared-messages.import": "Import",
+  "lib.shared-messages.importDevices": "Import Devices",
   "lib.shared-messages.integrations": "Integrations",
   "lib.shared-messages.joinAccept": "Join Accept",
   "lib.shared-messages.joinEUI": "JoinEUI",
