+++ pkg/webui/locales/xx.json
@@ -37,10 +37,6 @@
   "console.components.api-key-modal.index.buttonMessage": "X xxxx xxxxxx xxx xxx",
   "console.components.api-key-modal.index.grantedRights": "Xxxxxxx Xxxxxx",
   "console.components.api-key-modal.index.description": "Xxxx XXX Xxx xxx xxxx xxxxxxx xxxxxxxxxxxx.\nXxxx: Xxxxx xxxxxxx xxxx xxxxxx, xxx xxxxx xx xxx xxx xxxxxx xxxx xxx xx xxxxxxxxxx xxxxxxx.\nXxxx xxxx xx xxxx xxx xxxxx xx xx x xxxx xxxxx xxx.",
-  "console.components.device-bulk-create-form.index.fileImport": "Xxxx Xxxxxx",
-  "console.components.device-bulk-create-form.index.file": "Xxxx",
-  "console.components.device-bulk-create-form.index.createDevices": "Xxxxxx Xxxxxxx",
-  "console.components.device-bulk-create-form.index.selectAFile": "Xxxxxx xxxxxx x xxxxxxxx xxxx",
   "console.components.device-data-form.messages.abp": "Xxxxxxxxxx Xx Xxxxxxxxxxxxxxx (XXX)",
   "console.components.device-data-form.messages.activationMode": "Xxxxxxxxxx Xxxx",
   "console.components.device-data-form.messages.activationSettings": "Xxxxxxxxxx Xxxxxxxx",
@@ -73,6 +69,10 @@
   "console.components.device-data-form.messages.validate16": "Xxxx xxxxx xxxxx xx xx xxxxxxx xx xxxxxxxxxx xxxx",
   "console.components.device-data-form.messages.validate32": "Xxxx xxxxx xxxxx xx xx xxxxxxx xx xxxxxxxxxx xxxx",
   "console.components.device-data-form.messages.validate8": "Xxxx xxxxx xxxxx xx xx xxxxxxx x xxxxxxxxxx xxxx",
+  "console.components.device-import-form.index.fileImport": "Xxxx Xxxxxx",
+  "console.components.device-import-form.index.file": "Xxxx",
+  "console.components.device-import-form.index.createDevices": "Xxxxxx Xxxxxxx",
+  "console.components.device-import-form.index.selectAFile": "Xxxxxx xxxxxx x xxxxxxxx xxxx",
   "console.components.gateway-data-form.index.enforced": "Xxxxxxxx",
   "console.components.gateway-data-form.index.dutyCycle": "Xxxx Xxxxx",
   "console.components.gateway-data-form.index.gatewayIdPlaceholder": "xx-xxx-xxxxxxx",
@@ -131,14 +131,14 @@
   "console.containers.collaborators-table.index.id": "Xxxx / Xxxxxxxxxxxx XX",
   "console.containers.dev-addr-input.dev-addr-input.generate": "Xxxxxxxx Xxxxxx Xxxxxxx",
   "console.containers.dev-addr-input.index.devAddrFetchingFailure": "Xxxxx xxx xxxxxxxx xxxxxx xxxxxxx",
-  "console.containers.device-bulk-creator.index.proceed": "Xxxxxxx",
-  "console.containers.device-bulk-creator.index.retry": "Xxxxx",
-  "console.containers.device-bulk-creator.index.converting": "Xxxxxxxxxx Xxxxxxxxxx",
-  "console.containers.device-bulk-creator.index.creating": "Xxxxxxxx xxxxxxxx",
-  "console.containers.device-bulk-creator.index.operationInProgress": "Xxxxxxxxx xx xxxxxxxx",
-  "console.containers.device-bulk-creator.index.operationHalted": "Xxxxxxxxx xxxxxx",
-  "console.containers.device-bulk-creator.index.operationFinished": "Xxxxxxxxx xxxxxxxx",
-  "console.containers.device-bulk-creator.index.errorTitle": "Xxxxx xxx xxxxxxxx xxxxxxxxx",
+  "console.containers.device-importer.index.proceed": "Xxxxxxx",
+  "console.containers.device-importer.index.retry": "Xxxxx",
+  "console.containers.device-importer.index.converting": "Xxxxxxxxxx Xxxxxxxxxx",
+  "console.containers.device-importer.index.creating": "Xxxxxxxx xxxxxxxx",
+  "console.containers.device-importer.index.operationInProgress": "Xxxxxxxxx xx xxxxxxxx",
+  "console.containers.device-importer.index.operationHalted": "Xxxxxxxxx xxxxxx",
+  "console.containers.device-importer.index.operationFinished": "Xxxxxxxxx xxxxxxxx",
+  "console.containers.device-importer.index.errorTitle": "Xxxxx xxx xxxxxxxx xxxxxxxxx",
   "console.containers.device-template-format-select.index.title": "Xxxxxx",
   "console.containers.device-template-format-select.index.warning": "Xxxxx xxx xxxxxxxx xxx xxxx xx xxxxxxxxx xxxxxx xxxxxxxx xxxxxxx",
   "console.containers.freq-plans-select.index.title": "Xxxxxxxxx Xxxx",
@@ -266,19 +266,17 @@
   "lib.shared-messages.approve": "Xxxxxxx",
   "lib.shared-messages.appSKey": "XxxXXxx",
   "lib.shared-messages.brand": "Xxxxx",
-  "lib.shared-messages.bulk": "Xxxx",
-  "lib.shared-messages.bulkCreation": "Xxxx Xxxxxxxx",
   "lib.shared-messages.cancel": "Xxxxxx",
   "lib.shared-messages.changeLocation": "Xxxxxx xxxxxxxx xxxxxxxx",
   "lib.shared-messages.changePassword": "Xxxxxx Xxxxxxxx",
   "lib.shared-messages.clear": "Xxxxx",
+  "lib.shared-messages.collaborator": "Xxxxxxxxxxxx",
   "lib.shared-messages.collaboratorDeleteSuccess": "Xxxxxxxxxxxx xxxxxxx xxxxxxxxxxxx",
   "lib.shared-messages.collaboratorEdit": "Xxxx {collaboratorId}",
   "lib.shared-messages.collaboratorEditRights": "Xxxx xxxxxx xx {collaboratorId}",
   "lib.shared-messages.collaboratorId": "Xxxxxxxxxxxx XX",
   "lib.shared-messages.collaboratorModalWarning": "Xxx xxx xxxx xxx xxxx xx xxxxxx {collaboratorId} xx x xxxxxxxxxxxx?",
   "lib.shared-messages.collaboratorRemove": "Xxxxxxxxxxxx Xxxxxx",
-  "lib.shared-messages.collaborator": "Xxxxxxxxxxxx",
   "lib.shared-messages.collaborators": "Xxxxxxxxxxxxx",
   "lib.shared-messages.collaboratorUpdateSuccess": "Xxxxxxxxxxxx xxxxxxx xxxxxxxxxxxx xxxxxx",
   "lib.shared-messages.componentApplicationServer": "Xxxxxxxxxxx Xxxxxx",
@@ -330,6 +328,8 @@
   "lib.shared-messages.hardware": "Xxxxxxxx",
   "lib.shared-messages.hardwareVersion": "Xxxxxxxx xxxxxxx",
   "lib.shared-messages.id": "XX",
+  "lib.shared-messages.import": "Xxxxxx",
+  "lib.shared-messages.importDevices": "Xxxxxx Xxxxxxx",
   "lib.shared-messages.integrations": "Xxxxxxxxxxxx",
   "lib.shared-messages.joinAccept": "Xxxx Xxxxxx",
   "lib.shared-messages.joinEUI": "XxxxXXX",
